refactor(tui): colocate GistDetail mouse dispatch into screens/detail.rs
First of a 2-PR series colocating mouse dispatch the way keyboard
dispatch already is (issue #310). Moves the GistDetail-only pieces
out of keys.rs into screens/detail.rs's existing impl AppState block:

- apply_navigation's Screen::GistDetail arm -> apply_navigation_detail
  (plus its detail_nav helper, only ever called from that arm)
- click_select's Screen::GistDetail arm -> click_select_detail
- click_detail_tab and click_comments_load_older, moved wholesale
  (names unchanged; keys.rs's handle_mouse still calls them via
  self.<name>(...), so the call sites don't change)

keys.rs's apply_navigation/click_select shrink to one-line delegation
for this arm, matching how wheel_step already delegates. NavAction
widens from private to pub(crate) so detail.rs can name it in the new
methods' signatures; it stays in keys.rs alongside its producer,
nav_action().

GistDetail is the largest/riskiest slice (~80 lines across the four
pieces), so it ships on its own; the remaining screens (List, Gists,
Pins, Revisions, Help, Config) follow in one bundled PR. The
Diff/Preview/Confirm shared scroll arm stays in keys.rs -- identical
behavior across all three screens, so splitting it into three
delegating stubs would fight colocation rather than serve it.

No behavior change -- moved bodies are byte-identical, only their
file/visibility changed.

Part of #310
